+++ VSCodeExtensions/Public/Find-VSCodeExtension.ps1
@@ -1,8 +1,6 @@
 # .ExternalHelp VSCodeExtensions-Help.xml
 function Find-VSCodeExtension
 {
-<#
-#>
     [CmdletBinding(DefaultParameterSetName="ExtensionName")]
     [OutputType([PSCustomObject])]
     param
@@ -84,9 +82,11 @@ function Find-VSCodeExtension
         {
             foreach ($Extension in $Extensions)            
             {
+                $Extension | Add-Member -MemberType NoteProperty -Name Installs -Value (($Extension.statistics | Where-Object { $_.statisticName -eq 'Install' }).value)
                 $Extension | Add-Member -MemberType NoteProperty -Name publisherName -Value $Extension.publisher.publisherName
                 $Extension | Add-Member -MemberType NoteProperty -Name FullName -Value ($Extension.publisher.publisherName + '.' + $Extension.ExtensionName)
                 $Extension | Add-Member -MemberType NoteProperty -Name Version -Value $Extension.Versions[0].version
+                $Extension | Add-Member -MemberType NoteProperty -Name assetUri -Value $Extension.versions[0].assetUri
                 $Extension
             }
         }