+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 ﻿# VSCodeExtensions Release History
 
+## 0.1.0    7-17-2016
+Imporved the handling of VSIX files.
+Commands
+- Find-VSCodeExtension
+- Get-VSCodeExtension
+- Get-VSCodeExtensionStats
+- Get-VSCodePublisherStats
+- Install-VSCodeExtension
+- Uninstall-VSCodeExtension
+- Update-VSCodeExtension
+
+
 ## 0.0.1
 Initial Release